PlotInteractionEffectsContinuous

Plots the interaction effects between continuous variables in a dataset.

Usage

PlotInteractionEffectsContinuous(
  Data,
  interVar = NULL,
  xVars = NULL,
  yVars = NULL,
  covars = NULL,
  Relabel = TRUE,
  Ordinal = FALSE
)

Arguments

Data

The dataset to be analyzed

interVar

The variable of interest for the interaction effect

xVars

The variables to be plotted on the x-axis

yVars

The variables to be plotted on the y-axis

covars

Covariates to be included in the analysis

Relabel

A logical value indicating whether to relabel the variables with their corresponding labels

Ordinal

A logical value indicating whether the variables are ordinal

Value

A list containing the results of the analysis, including the p-values, significance levels, and the ggplot object.